spi: geni-msm: Remove geni-spi compatible field
The geni-spi compatible field is being removed from the newly added
driver file for the SA8255P platform. This change is necessary to
prevent potential conflicts with multiple drivers that could arise due
to the presence of this field.

By removing it, we aim to ensure better compatibility and stability
for the driver.

Change-Id: Ia2f950192e9f992548f13a476d30cbf48dcfa9ea
Signed-off-by: Praveen Talari <quic_ptalari@quicinc.com>