i2c: geni-msm: Remove geni-i2c compatible field
The geni-i2c compatible field is being removed from the newly added
driver file for the SA8255P platform. This change is necessary to
prevent potential conflicts with multiple drivers that could arise due
to the presence of this field.

By removing it, we aim to ensure better compatibility and stability
for the driver.

Change-Id: Ic390ae4b2b7037144b5a5739c1b29dd6d8c1c9c9
Signed-off-by: Praveen Talari <quic_ptalari@quicinc.com>