usb: dwc3: dwc3-msm-core: Add support for passing orientation from DeviceTree
In targets where a PD solution is not available to notify whether the
connected orientation is CC1 or CC2, the default value is generally
used. However, on some boards, it is ideal to control the orientation
from the DeviceTree to address issues like reverse orientation.

This patch enables the feature to read the orientation passed from the
DeviceTree and configure the hardware accordingly.

Change-Id: I14af0ac6e8b8c77a7105f82f0aa43a4eb4040378
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>