remoteproc: qcom: pas: Add late attach support for subsystems
Subsystems can be brought out of reset by entities such as
bootloaders. Before attaching such subsystems, it is important to
check the state of the subsystem. This patch adds support to attach
to a subsystem by ensuring that the subsystem is in a sane state by
reading SMP2P bits and pinging the subsystem.

Change-Id: I0a2ef8831104503e12278b56e10dc2d79f50d079
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>
Signed-off-by: David Gantman <quic_dgantman@quicinc.com>
Signed-off-by: Auditya Bhattaram <quic_audityab@quicinc.com>
Signed-off-by: Navya Vemula <quic_nvemula@quicinc.com>